[FIX] Incorrect column name
